refactor: 경매글 생성 시, 경매 서비스에 초기 데이터 메시지 받아서 round_info 초기 셋팅 및 경매 시작과 경매 마감 스케줄러 처리(#187)

+++ src/main/java/com/skyhorsemanpower/auction/kafka/KafkaConsumerCluster.java
@@ -36,8 +36,8 @@ public class KafkaConsumerCluster {
                 .auctionUuid(message.get("auctionUuid").toString())
                 .startPrice(new BigDecimal(message.get("startPrice").toString()))
                 .numberOfEventParticipants((Integer) message.get("numberOfEventParticipants"))
-                .auctionStartTime((Long) message.get("auctionStartTime"))
-                .auctionEndTime((Long) message.get("auctionEndTime"))
+                .auctionStartTime(((Long) message.get("auctionStartTime")).intValue())
+                .auctionEndTime(((Long) message.get("auctionEndTime")).intValue())
                 .incrementUnit(new BigDecimal(message.get("incrementUnit").toString()))
                 .build();
         log.info("InitialAuctionDto >>> {}", initialAuctionDto.toString());
